``` feat: implement capture image using camera intent
This commit implements image capture functionality using camera intent.
It utilizes the `ActivityResultContracts.TakePicture` contract
``````
 to capture images.
It also includes logic to create a unique file path using `createImageUri` and `generateFilename` functions.
This implementation removes the use of `CameraXViewfinder` and `CameraPreviewViewModel`, because it's not stable yet, for android 11 bellow.
The camera permission is requested using `rememberPermission State`.
```