update wallet balance when saving a transaction
This commit updates the `SaveTransactionUseCaseImpl` to update the wallet balance when a transaction is saved.

The `SaveTransactionUseCaseImpl` now calls the `walletRepository.update()` method to update the wallet balance based on the transaction type and amount.

If the transaction type is `Income`, the amount is added to the wallet balance.

If the transaction type is `Expense`, the amount is subtracted from the wallet balance.