refactor: move wallet dropdown component to ui module
Move the wallet dropdown component from the `addTransaction` feature module to the `ui` module to promote reusability.

This change involves moving the `SelectWalletDropdownMenuView`, `SelectWalletDropdownMenuState`, and `rememberSelectWalletDropdownMenuState` components to the `ui` module.

Additionally, the database schema is updated to include a `GoalTransactionHistoryTable` to store goal transactions. A new `GoalTransactionRepository` is introduced to handle saving goal transactions.

The `GoalViewModel` is updated to use the new wallet dropdown component and to load wallets and the selected wallet.

Finally, the `DialogAddSavingView` is updated to use the new wallet dropdown component and to add a note field.